Better Blackwater (Budget Goodsprings PR)
## About The Pull Request

Adds Blackwater, an alternative to Blackwater (186) in the form of a Goodsprings-inspired western frontier town.
### Map preview
<details>
  <summary>Images</summary>
   
![image](https://user-images.githubusercontent.com/72716882/218872600-bbcdecab-8a4f-40ce-bc33-931f0cc3d89f.png)
![image](https://user-images.githubusercontent.com/72716882/218872697-08eb8aff-eeaa-4b43-8f30-1aeb46e9ccc3.png)
![image](https://cdn.discordapp.com/attachments/1014004595950895106/1075184036370206830/image.png)


</details>
Contains:
Heavily abstracted roles. Only "Blackwater Settler" and "Blackwater Deputy" (3 slots), both located in the wasteland catagory and offering the same footprint as rustwalker. The actual individual roles in the town are either filled by the deputies choosing amongst themselves who would wear the Sheriff's badge, or settlers entering a building and placing a lock on the door.

Locations:

- Sheriff's office (armory consists of a cowboy repeater, a caravan shotgun, a 223 pistol with rubbers in the attic, and the .357 revolvers they spawn with.)
- Saloon with kitchen and adjacent garden/brahmin field
- Doc's house (has a fridge with some bloodbags, a set of organs, food, a surgical dining table, a primative surgical kit, books, and a car battery in the car outside)
- General store (contains a workbench, metal bench, lathe, ammo handloading tool (no ammo bench), and some starting guns)
- 2 inhabited houses and a tent.
- Vendotron point with a C.A.M.P and wasteland medical vending machines, in case the doc was thoughtful enough to get himself killed after reading the books.
- Abandoned, infested schoolhouse
- Some caves with a tier 2 loot spawn.

## Why It's Good For The Game

The gripes with Kebab that inspired Hangyul's Blackwater were regarding the lack of townie roleplay, however, the base and faction proposed in his PR only provided a small, isolated commune walled off from outsiders. This doesn't provide a roleplay hub that was lost with previous iterations of the town. Anyone outside the blackwater pseudo-tribe can't enter the "town", thus not providing anything outside of the clique. Nonhostile factions should be interactable with outsiders to actually contribute something to the round, so I mapped in a small goodsprings-styled town with a handful of buildings and minimal roles. 

This frontier town provides a unique, western theme that contrasts greatly with the immoral, down and dirty city with it's rival factions and pickpocketing scabs. 

## Pre-Merge Checklist
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.


## Changelog

:cl:
add: Adds blackwater
/:cl: